+++ data-raw/dat.michael2013.R
@@ -1,3 +1,0 @@
-dat.michael2013 <- read.csv("data-raw/dat.michael2013.csv")
-
-save(dat.michael2013, file="data/dat.michael.rda", ascii=FALSE)